fix: NUMERIC type mismatch — use rust_decimal for g_earned_lifetime + price_g; disable statement cache for PgBouncer compat

+++ apps/api/src/handlers/battles.rs
@@ -132,7 +132,7 @@ pub async fn fight_bot(
         .bind(wins)
         .bind(losses)
         .bind(rank)
-        .bind(g_awarded as f64)
+        .bind(g_awarded)
         .bind(now)
         .bind(&wallet)
         .execute(&state.db)
